+++ osmoutils/accum/errors.go
@@ -2,14 +2,12 @@ package accum
 
 import (
 	"fmt"
-
-	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
 type NoPositionError struct {
-	Address sdk.AccAddress
+	Name string
 }
 
 func (e NoPositionError) Error() string {
-	return fmt.Sprintf("no position found for address (%s)", e.Address)
+	return fmt.Sprintf("no position found for address (%s)", e.Name)
 }